% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_labels.R
\name{alter_labels}
\alias{alter_labels}
\title{Alter an incidence plot labels}
\usage{
alter_labels(plot, angle = 90)
}
\arguments{
\item{plot}{A ggplot2 incidence graph.}

\item{angle}{Angle to rotate x-axis labels (default 90 degrees)}
}
\description{
Alter an incidence plot labels
}
