% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.data.frame.incidence2}
\alias{as.data.frame.incidence2}
\title{Convert incident object to a data frame}
\usage{
\method{as.data.frame}{incidence2}(x, row.names, optional, ...)
}
\arguments{
\item{x}{\verb{<incidence2>} object.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{...}{Not used.}
}
\description{
Convert incident object to a data frame
}
\examples{
dat <- data.frame(
    dates = Sys.Date() + 1:100,
    names = rep(c("Jo", "John"), 5)
)

dat <- incidence(dat, date_index = "dates", groups = "names")
as.data.frame(dat)

}
