% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.incidence}
\title{Convert incident object to a tibble}
\usage{
\method{as_tibble}{incidence}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{...}{Not used.}
}
\description{
Convert incident object to a tibble
}
\examples{
dat <- data.frame(dates = Sys.Date() + 1:100,
                  names = rep(c("Jo", "John"), 5))

dat <- incidence(dat, date_index = dates, groups = names)
as_tibble(dat)

}
