% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.data.frame.incidence}
\alias{as.data.frame.incidence}
\alias{as_tibble.incidence}
\title{Conversion of incidence objects}
\usage{
\method{as.data.frame}{incidence}(x, ...)

as_tibble.incidence(x, ...)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{...}{Not used.}
}
\description{
These functions convert \code{incidence} objects into other classes.
}
\seealso{
the \code{\link[=incidence]{incidence()}} function to generate the 'incidence' objects.
}
