% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\name{cumulate-deprecated}
\alias{cumulate-deprecated}
\alias{cumulate}
\title{Compute cumulative 'incidence'}
\usage{
cumulate(x)
}
\arguments{
\item{x}{An incidence object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{cumulate} has been deprecated to avoid encouraging erroneous regression
on cumulative counts.
}
\keyword{internal}
