% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peak.R
\name{find_peak}
\alias{find_peak}
\title{Find the peak date of an incidence curve}
\usage{
find_peak(x, pool = TRUE)
}
\arguments{
\item{x}{An \code{incidence} object.}

\item{pool}{If \code{TRUE} (default), any groups will be pooled before finding
a peak. If \code{FALSE}, separate peaks will be found for each group.}
}
\value{
The date of the (first) highest incidence in the data.
}
\description{
This function can be used to find the peak of an epidemic curve stored as an
\code{incidence} object.
}
\seealso{
\code{\link[=estimate_peak]{estimate_peak()}} for bootstrap estimates of the peak time
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}, Zhian N. Kamvar
\email{zkamvar@gmail.com}
}
