% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interval_duration.R
\name{get_interval_duration.incidence2}
\alias{get_interval_duration.incidence2}
\title{Days covered by the date index of an incidence object}
\usage{
\method{get_interval_duration}{incidence2}(x, ...)
}
\arguments{
\item{x}{An incidence object.}

\item{...}{Not currently used.}
}
\value{
The number of days covered from the start to end dates for each element in
the date_index.
}
\description{
Utility function for returning the number of days covered by an element
of the incidence objects date_index column.
}
\examples{
\dontshow{.old <- data.table::setDTthreads(2)}
if (requireNamespace("outbreaks", quietly = TRUE)) {
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    i <- incidence(dat, "date_of_onset", interval = "isoweek")
    get_interval_duration(i)
}
\dontshow{data.table::setDTthreads(.old)}


}
