% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{incidence2-defunct}
\alias{incidence2-defunct}
\alias{new_incidence}
\alias{validate_incidence}
\alias{build_incidence}
\alias{get_n}
\alias{get_interval}
\alias{get_timespan}
\alias{facet_plot}
\title{Functions now defunct in package incidence2}
\usage{
new_incidence(
  x,
  date,
  groups = NULL,
  counts,
  measurements = NULL,
  validate = TRUE
)

validate_incidence(x)

build_incidence(
  x,
  date_index,
  groups = NULL,
  counts = NULL,
  na_as_group = TRUE,
  FUN = identity,
  args = list()
)

get_n(x)

get_interval(x, ...)

get_timespan(x, ...)

facet_plot(x, ...)
}
\description{
These functions are now defunct.
}
\keyword{internal}
