% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence2-package.R
\docType{package}
\name{incidence2-package}
\alias{incidence2}
\alias{incidence2-package}
\title{incidence2: Compute, Handle and Plot Incidence of Dated Events}
\description{
Provides functions and classes to compute, handle and visualise incidence from dated events for a defined time interval. Dates can be provided in various standard formats. The class 'incidence2' is used to store computed incidence and can be easily manipulated, subsetted, and plotted. This package is part of the RECON (\url{https://www.repidemicsconsortium.org/}) toolkit for outbreak analysis (\url{https://www.reconverse.org}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reconverse/incidence2}
  \item Report bugs at \url{https://github.com/reconverse/incidence2/issues}
}

}
\author{
\strong{Maintainer}: Thibaut Jombart \email{thibautjombart@gmail.com} [contributor]

Authors:
\itemize{
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})
}

}
\keyword{internal}
