% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vibrant}
\alias{vibrant}
\alias{palettes}
\alias{muted}
\title{Color palettes used in incidence}
\usage{
vibrant(n)

muted(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in incidence. The palettes come from
https://personal.sron.nl/~pault/#sec:qualitative and exclude \code{grey}, which
is reserved for missing data.
}
\examples{
vibrant(5)
muted(10)
}
