% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/scale_x_incidence.R
\name{plot.incidence}
\alias{plot.incidence}
\alias{scale_x_incidence}
\alias{make_breaks}
\title{Plot an incidence object}
\usage{
\method{plot}{incidence}(
  x,
  color = "black",
  col_pal = incidence_pal1,
  alpha = 0.7,
  border = NA,
  xlab = "",
  ylab = NULL,
  labels_week = has_weeks(x),
  labels_iso = has_isoweeks(x),
  show_cases = FALSE,
  n_breaks = 6,
  facet = TRUE,
  stack = c("none", "color", "dodge"),
  ...
)

scale_x_incidence(x, n_breaks = 6, labels_week = TRUE, ...)

make_breaks(x, n_breaks = 6L, labels_week = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}},
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}}, or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, depending
on how the \verb{$date} element is stored in the incidence object.}
}
\description{
Plot an incidence object
}
