% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup.R
\name{pool}
\alias{pool}
\title{Pool 'incidence' objects}
\usage{
pool(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
Pool was a function from the original incidence package that has now been
deprecated in favour of \code{\link[=regroup]{regroup()}}
}
\keyword{internal}
