% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.incidence2}
\alias{print.incidence2}
\title{Print an incidence object.}
\usage{
\method{print}{incidence2}(x, ...)
}
\arguments{
\item{x}{An 'incidence2' object.}

\item{...}{Not used.}
}
\description{
Print an incidence object.
}
