% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_incidence}
\alias{print_incidence}
\alias{print.incidence_df}
\alias{format.incidence_df}
\title{Print an incidence object.}
\usage{
\method{print}{incidence_df}(x, ...)

\method{format}{incidence_df}(x, ...)
}
\arguments{
\item{x}{An 'incidence' object.}

\item{...}{Additional arguments passed through to the tibble format method.}
}
\description{
Print an incidence object.
}
