% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R, R/as_tibble.R, R/mutate.R,
%   R/nest.R, R/summarise.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.table}
\alias{as_tibble}
\alias{mutate}
\alias{nest}
\alias{summarise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{nest}}}
}}

