% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.incidence}
\alias{summary.incidence}
\title{Summary of an incidence object}
\usage{
\method{summary}{incidence}(object, ...)
}
\arguments{
\item{object}{An 'incidence' object.}

\item{...}{Not used.}
}
\value{
object (invisibly).
}
\description{
Summary of an incidence object
}
\examples{

data(ebola_sim_clean, package = "outbreaks")
dat <- ebola_sim_clean$linelist
inci <- incidence(dat, "date_of_onset", groups = c("gender", "hospital"))
summary(inci)

}
