% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.incidence2}
\alias{summary.incidence2}
\title{Summary of a given incidence object}
\usage{
\method{summary}{incidence2}(object, ...)
}
\arguments{
\item{object}{An 'incidence' object.}

\item{...}{Not used.}
}
\value{
object (invisibly).
}
\description{
Summary of a given incidence object
}
