% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.incidence_df}
\alias{summary.incidence_df}
\title{Summary of an incidence object}
\usage{
\method{summary}{incidence_df}(object, ...)
}
\arguments{
\item{object}{An 'incidence' object.}

\item{...}{Not used.}
}
\value{
object (invisibly).
}
\description{
Summary of an incidence object
}
