% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/get_interval.R, R/get_n.R,
%   R/get_timespan.R
\name{accessors}
\alias{accessors}
\alias{get_count_name}
\alias{get_count_name.default}
\alias{get_count_name.incidence}
\alias{get_date_group_names}
\alias{get_date_name}
\alias{get_date_group_names.default}
\alias{get_date_group_names.incidence}
\alias{get_date_name.incidence}
\alias{get_date_name.default}
\alias{get_group_names}
\alias{get_group_names.default}
\alias{get_group_names.incidence}
\alias{get_interval}
\alias{get_interval.default}
\alias{get_interval.incidence}
\alias{get_n}
\alias{get_n.default}
\alias{get_n.incidence}
\alias{get_timespan}
\alias{get_timespan.default}
\alias{get_timespan.incidence}
\title{Access various elements of an incidence object}
\usage{
get_count_name(x, ...)

\method{get_count_name}{default}(x, ...)

\method{get_count_name}{incidence}(x, ...)

get_date_group_names(x, ...)

\method{get_date_group_names}{default}(x, ...)

\method{get_date_group_names}{incidence}(x, ...)

get_date_name(x, ...)

\method{get_date_name}{default}(x, ...)

\method{get_date_name}{incidence}(x, ...)

get_group_names(x, ...)

\method{get_group_names}{default}(x, ...)

\method{get_group_names}{incidence}(x, ...)

get_interval(x, ...)

\method{get_interval}{default}(x, ...)

\method{get_interval}{incidence}(x, integer = FALSE, ...)

get_n(x)

\method{get_n}{default}(x)

\method{get_n}{incidence}(x)

get_timespan(x, ...)

\method{get_timespan}{default}(x, ...)

\method{get_timespan}{incidence}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{...}{Not used.}

\item{integer}{When \code{TRUE}, the interval will be converted to an
integer vector if it is stored as a character in the incidence object.}
}
\value{
\itemize{
\item \code{get_count_name()}: The count variable of x.
}

\itemize{
\item \code{get_date_group_names()}: a character vector of the date variables of x.
}

\itemize{
\item \code{get_date_name()}: a character vector of the date variables of x.
}

\itemize{
\item \code{get_group_names()}: a character vector of the group variables of x or
NULL if none are present.
}

\itemize{
\item \code{get_interval()}: if \code{integer = TRUE}, an integer vector, otherwise the
character value of the \code{interval}
}

\itemize{
\item \code{get_n()} The total number of cases stored in the object
}

\itemize{
\item \code{get_timespan()}: an \code{integer} denoting the timespan in days represented
by the incidence object.
}
}
\description{
Access various elements of an incidence object
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    i <- incidence(dat,
                   date_index = date_of_onset,
                   groups = c(gender, hospital))

    get_count_name(i)

    get_group_names(i)

    get_date_name(i)

    get_date_group_names(i)

    get_interval(i)

    get_n(i)

    get_timespan(i)
  })
}

}
