% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_incidence.R
\name{as_incidence}
\alias{as_incidence}
\alias{as_incidence.default}
\alias{as_incidence.data.frame}
\title{Coerce to incidence}
\usage{
as_incidence(x, ...)

\method{as_incidence}{default}(x, ...)

\method{as_incidence}{data.frame}(
  x,
  date_index,
  groups = NULL,
  counts = NULL,
  na_as_group = TRUE,
  FUN = identity,
  args = list(),
  ...
)
}
\arguments{
\item{x}{An object to converse to \code{incidence}.}

\item{...}{Further arguments passed to or from other methods.}

\item{date_index}{The time index(es) of the given data. Multiple inputs only
make sense when x is a linelist, and in this situation, to avoid ambiguity,
the vector must be named.  These names will be used for the resultant count
columns.}

\item{groups}{An optional vector giving the names of the groups of
observations for which incidence should be grouped.}

\item{counts}{The count variables of the given data.  If NULL (default) the
data is taken to be a linelist of individual observations.}

\item{na_as_group}{A logical value indicating if missing group values (NA)
should treated as a separate category (\code{TRUE}) or removed from
consideration (\code{FALSE}). Defaults to \code{TRUE}.}

\item{FUN}{Function applied to the \code{dates_index} vectors before grouping. The
first argument of \code{FUN} must work with a \code{dates_index} vector. Defaults to
the identity function.}

\item{args}{List of additional arguments passed to FUN.}
}
\description{
\code{as_incidence()} coerced an object to an incidence of events.
}
