% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_incidence}
\alias{as_incidence}
\alias{as_incidence.default}
\alias{as_incidence.incidence2}
\title{Coerce to an incidence object}
\usage{
as_incidence(x, ...)

\method{as_incidence}{default}(x, ...)

\method{as_incidence}{incidence2}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
An \verb{<incidence2>} object.
}
\description{
Generic for coercion to an \verb{<incidence2>} object.
}
