% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\name{cumulate}
\alias{cumulate}
\alias{cumulate.default}
\alias{cumulate.incidence}
\title{Compute cumulative 'incidence'}
\usage{
cumulate(x)

\method{cumulate}{default}(x)

\method{cumulate}{incidence}(x)
}
\arguments{
\item{x}{An incidence object.}
}
\description{
\code{cumulate} is an S3 generic to compute cumulative numbers, with methods
for different types of objects:
\itemize{
\item default method is a wrapper for \code{cumsum}
\item \code{incidence} objects: computes cumulative incidence over time
}
}
\seealso{
The \code{\link[=incidence]{incidence()}} function to generate the 'incidence'
objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
