% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{deprecated-accessors}
\alias{deprecated-accessors}
\alias{get_counts_name}
\alias{get_date_group_names}
\title{Deprecated accessor functions}
\usage{
get_counts_name(x, ...)

get_date_group_names(x, ...)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{...}{Not used.}
}
\value{
\itemize{
\item \code{get_counts_name()}: Same as \code{get_count_names()}.
}

\itemize{
\item \code{get_date_group_names()}: Same as \code{get_dates_name()}
}
}
\description{
These accessor functions have now been deprecated in favour of better
named alternatives but are left available for backwards compatibility.
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    i <- incidence(dat,
                   date_index = date_of_onset,
                   groups = c(gender, hospital))

    get_counts_name(i)

    get_date_group_names(i)
  })
}

}
\keyword{internal}
