% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Compute the incidence of events}
\usage{
incidence(
  x,
  date_index,
  groups = NULL,
  counts = NULL,
  count_names_to = "count_variable",
  count_values_to = "count",
  rm_na_dates = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame object representing a linelist or pre-aggregated dataset.}

\item{date_index}{\verb{[character]}

The time index(es) of the given data.  This should be the name(s)
corresponding to the desired date column(s) in x. Multiple indices only make
sense when  \code{x} is a linelist; in this situation the vector must be named and
these names will be used for the resultant count variables.}

\item{groups}{\verb{[character]}

An optional vector giving the names of the groups of observations for which
incidence should be grouped.}

\item{counts}{\verb{[character]}

The count variables of the given data.  If NULL (default) the data is taken
to be a linelist of individual observations.}

\item{count_names_to}{\verb{[character]}

The column to create which will store the \code{counts} column names provided that
\code{counts} is not NULL.}

\item{count_values_to}{\verb{[character]}

The name of the column to store the resultant count values in.}

\item{rm_na_dates}{\verb{[logical]}

Should \code{NA} dates be removed prior to aggregation?}

\item{...}{Not currently used}
}
\value{
An object of class \verb{[incidence, data.frame]}.
}
\description{
\code{incidence()} calculates event the \emph{incidence} of different events across
specified time periods and groupings.
}
\details{
\verb{<incidence>} objects are a sub class of data frame with some
additional invariants. That is, an \verb{<incidence>} object must:
\itemize{
\item have one column representing the date index (this does not need to be a
\code{date} object but must have an inherent ordering over time);
\item have one column representing the count variable (i.e. what is being
counted) and one variable representing the associated count;
\item have zero or more columns representing groups;
\item not have duplicated rows with regards to the date and group variables.
}
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
\dontshow{withAutoprint(\{}
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    incidence(dat, "date_of_onset")
    incidence(dat, "date_of_onset", groups = c("gender", "hospital"))
\dontshow{\})}
}


}
