% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R, R/print.R, R/summary.R
\name{incidence}
\alias{incidence}
\alias{incidence.default}
\alias{incidence.Date}
\alias{incidence.character}
\alias{incidence.integer}
\alias{incidence.numeric}
\alias{incidence.POSIXt}
\alias{print.incidence}
\alias{summary.incidence}
\title{Create an incidence object}
\usage{
incidence(x, date_index, interval = 1L, ...)

\method{incidence}{default}(x, date_index, interval = 1L, ...)

\method{incidence}{Date}(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  ...
)

\method{incidence}{character}(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  ...
)

\method{incidence}{integer}(
  x,
  date_index,
  interval = 1L,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  ...
)

\method{incidence}{numeric}(
  x,
  date_index,
  interval = 1L,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  ...
)

\method{incidence}{POSIXt}(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  ...
)

\method{print}{incidence}(x, ...)

\method{summary}{incidence}(object, ...)
}
\arguments{
\item{x}{An 'incidence' object.}

\item{date_index}{The time index of the given data.  This should be the name,
with or without quotation, corresponding to a date column in x of the
class:  integer, numeric, Date, POSIXct, POSIXlt, and character. (See Note
about \code{numeric} and \code{character} formats)}

\item{interval}{An integer or character indicating the (fixed) size of the
time interval used for computing the incidence; defaults to 1 day. This can
also be a text string that corresponds to a valid date interval: day, week,
month, quarter, or year. (See Note).}

\item{...}{Additional arguments used by other methods.}

\item{standard}{(Only applicable where date_index references a Date object)
When \code{TRUE} (default) and the \code{interval} one of "week", "month", "quarter",
or "year", then this will cause the bins for the counts to start at the
beginning of the interval (See Note).}

\item{groups}{An optional vector giving the names of the groups of
observations for which incidence should be grouped.  This can be given with
or without quotation.}

\item{na_as_group}{A logical value indicating if missing group (NA) should be
treated as a separate group.}

\item{first_date, last_date}{optional first/last dates to be used. When
these are \code{NULL} (default), the dates from the first/last dates are taken
from the observations. If these dates are provided, the observations will be
trimmed to the range of [first_date, last_date].}

\item{object}{An 'incidence' object.}
}
\value{
An incidence object.
}
\description{
Create an incidence object

Print an incidence object.

Summary of a given incidence object
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  data(ebola_sim_clean, package = "outbreaks")
  dat <- ebola_sim_clean$linelist
  i1 <- incidence(dat,
    date_index = date_of_onset,
    interval = 1L,
    first_date = "2014-05-20",
    last_date = "2014-06-10",
    groups = c(hospital, gender)
  )
}
}
