% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Compute the incidence of events}
\usage{
incidence(
  x,
  date_index,
  groups = NULL,
  interval = 1L,
  na_as_group = TRUE,
  count = NULL,
  ...
)
}
\arguments{
\item{x}{A tibble or a data frame (see Note) representing a linelist.}

\item{date_index}{The time index of the given data.  This should be a name
corresponding to a date column in x of class:  integer, numeric, Date,
POSIXct, POSIXlt, and character. (See Note  about \code{numeric} and \code{character}
formats).}

\item{groups}{An optional vector giving the names of the groups of
observations for which incidence should be grouped.}

\item{interval}{An integer or character indicating the (fixed) size of the
time interval used for computing the incidence; defaults to 1 day. This can
also be a text string that corresponds to a valid date interval:
"yearweek", "yearmonth", "yearquarter" and also "day", "week", "month",
"quarter", or "year". (See Note).}

\item{na_as_group}{A logical value indicating if missing group values (NA)
should treated as a separate category (\code{TRUE}) or removed from
consideration (\code{FALSE}).}

\item{count}{The count variable of the given data.  If NULL (default) the
data is taken to be a linelist of individual observations.}

\item{...}{Additional arguments passed to the underlying date grouping
functions when the interval is "yearweek" or one of "day", "week", "month",
"quarter" or "year". In the case of "yearweek" you can optionally specify
an integer value \code{firstday} representing the day the week should start on;
1 (Monday) to 7 (Sunday).  If no value is specified then the default value
of 1 (Monday) is used which corresponds to the ISO 8601 definition of
weeks. For the other values you can optionally specify a value \code{firstdate}
which will be passed to the underlying function \code{\link[=as_period]{as_period()}} or, for
integer dates, \code{\link[=as_int_period]{as_int_period()}}.  This will cause the intervals created
to begin on that date.}

\item{firstday}{()}
}
\value{
An incidence2 object.  This is a subclass of tibble that represents
and aggregated count of observations grouped according to the specified
interval and, optionally, the given groups.  By default it will contain the
following columns:
\itemize{
\item \strong{date} / \strong{date_index}:  If the default interval of 1 day is used then
this will be the dates of the given observations and given the name
"date", otherwise, this will be values obtained from the specified date
grouping with column name "date_index" (See Interval specification below).
\item \strong{-groups-}: If specified, column(s) containing the categories of the
given groups.
\item \strong{count}: The aggregated observation count.
}
}
\description{
Compute the incidence of events
}
\note{
\subsection{Input data (\code{dates})}{
\itemize{
\item \strong{Decimal (numeric) dates}: will be truncated.
\item \strong{Character dates} should be in the unambiguous \code{yyyy-mm-dd} (ISO 8601)
format. Any other format will trigger an error.
}
}

\subsection{Interval specification (\code{interval})}{
\code{\link{incidence2}} includes generators for 5 different grouped date s3 classes.
Which one the \code{incidence()} function uses depends on the value of
\code{interval}. This can be specified as either an integer value or a more
standard specification such as "day", "week", "month", "quarter" or "year".
The format in this situation is similar to that used by \code{\link[=seq.Date]{seq.Date()}}
where these values can optionally be preceded by a (positive or negative)
integer and a space, or followed by "s".  When no prefix is given:
\itemize{
\item "week"    : uses the "yrwk" class (see \code{\link[=as_yrwk]{as_yrwk()}}).
\item "month"   : uses the "yrmon" class (see \code{\link[=as_yrmon]{as_yrmon()}}).
\item "quarter" : uses the "yrqtr" class (see \code{\link[=as_yrqtr]{as_yrqtr()}}).
\item "year"    : uses the "yr" class (see \code{\link[=as_yr]{as_yr()}}).
}

When a prefix is provided (e.g. 2 weeks) the output is an object of class
"period" (see \code{\link[=as_period]{as_period()}}).  Note that for the values "month",
"quarter" and "year" intervals are always chosen to start at the beginning
of the calendar equivalent.  If the input is an integer value the input is
treated as if it was specified in days (i.e. 2 and 2 days) produce the
same output.

The only interval values that do not produce these grouped classes are 1,
1L, "day" or "days" (both without prefix) are used.  In this situation the
returned object is of the standard "Date" class.
}
}
