% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\alias{keep_first}
\alias{keep_first.default}
\alias{keep_first.incidence2}
\alias{keep_first.yrwk}
\alias{keep_first.yrmon}
\alias{keep_first.yrqtr}
\alias{keep_first.yr}
\alias{keep_first.period}
\alias{keep_first.int_period}
\alias{keep_last}
\alias{keep_last.default}
\alias{keep_last.incidence2}
\alias{keep_last.yrwk}
\alias{keep_last.yrmon}
\alias{keep_last.yrqtr}
\alias{keep_last.yr}
\alias{keep_last.period}
\alias{keep_last.int_period}
\title{Keep first and last occurences}
\usage{
keep_first(x, n, ...)

\method{keep_first}{default}(x, n, ...)

\method{keep_first}{incidence2}(x, n, ...)

\method{keep_first}{yrwk}(x, n, ...)

\method{keep_first}{yrmon}(x, n, ...)

\method{keep_first}{yrqtr}(x, n, ...)

\method{keep_first}{yr}(x, n, ...)

\method{keep_first}{period}(x, n, ...)

\method{keep_first}{int_period}(x, n, ...)

keep_last(x, n, ...)

\method{keep_last}{default}(x, n, ...)

\method{keep_last}{incidence2}(x, n, ...)

\method{keep_last}{yrwk}(x, n, ...)

\method{keep_last}{yrmon}(x, n, ...)

\method{keep_last}{yrqtr}(x, n, ...)

\method{keep_last}{yr}(x, n, ...)

\method{keep_last}{period}(x, n, ...)

\method{keep_last}{int_period}(x, n, ...)
}
\arguments{
\item{x}{Object to filter.}

\item{n}{Number of entries to keep.}

\item{...}{Not currently used.}
}
\value{
The objected with the chosen entries.
}
\description{
\code{keep_first()} (\code{keep_last}) keeps the first (last) \code{n} entries to occur
by date ordering.
}
