% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\alias{keep_first}
\alias{keep_first.default}
\alias{keep_first.incidence2}
\alias{keep_last}
\alias{keep_last.default}
\alias{keep_last.incidence2}
\title{Keep first and last occurences}
\usage{
keep_first(x, n, ...)

\method{keep_first}{default}(x, n, ...)

\method{keep_first}{incidence2}(x, n, ...)

keep_last(x, n, ...)

\method{keep_last}{default}(x, n, ...)

\method{keep_last}{incidence2}(x, n, ...)
}
\arguments{
\item{x}{Object to filter.}

\item{n}{Number of entries.}

\item{...}{Not currently used.}
}
\value{
The objected with the chosen entries.
}
\description{
\code{keep_first()} (\code{keep_last}) keeps the first (last) \code{n} entries to occur
by date ordering.
}
