% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\alias{keep_first}
\alias{keep_last}
\title{Keep first and last occurences}
\usage{
keep_first(x, n)

keep_last(x, n)
}
\arguments{
\item{x}{\verb{[incidence]} object.}

\item{n}{\verb{[integer]}

Number of entries to keep.

\code{double} vectors will be converted via \code{as.integer(n)}.}
}
\value{
Incidence object with the chosen entries.
}
\description{
\code{keep_first()} (\code{keep_last}) keeps the first (last) \code{n} entries to occur by
date ordering.
}
\examples{

if (requireNamespace("outbreaks", quietly = TRUE)) {
\dontshow{withAutoprint(\{}
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    inci <- incidence(dat, "date_of_onset")
    keep_first(inci, 3)
    keep_last(inci, 3)
\dontshow{\})}
}
}
