% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{incidence_pal}
\alias{incidence_pal}
\alias{palettes}
\title{Color palette used in incidence}
\usage{
incidence_pal(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
This function creates the default color palette used in incidence. It is based
on the colour-blind friendly palette's of Paul Tol
(https://personal.sron.nl/~pault/#sec:qualitative)
}
\examples{

plot(1:4, cex=8, pch=20, col = incidence_pal(4),
     main = "palette: incidence_pal")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
