% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/scale_x_incidence.R
\name{plot.incidence}
\alias{plot.incidence}
\alias{facet_plot}
\alias{scale_x_incidence}
\alias{make_breaks}
\title{Plotting functions}
\usage{
\method{plot}{incidence}(
  x,
  group = TRUE,
  stack = TRUE,
  col_pal = vibrant,
  alpha = 0.7,
  color = NA,
  xlab = "",
  ylab = NULL,
  n_breaks = 6,
  show_cases = FALSE,
  border = "white",
  na_color = "grey",
  labels_week = has_weeks(x),
  ...
)

facet_plot(
  x,
  fill = NULL,
  col_pal = vibrant,
  alpha = 0.7,
  color = "white",
  xlab = "",
  ylab = NULL,
  n_breaks = 6,
  show_cases = FALSE,
  border = "white",
  labels_week = has_weeks(x),
  na_color = "grey",
  legend = TRUE,
  ...
)

scale_x_incidence(x, n_breaks = 6, labels_week = TRUE, ...)

make_breaks(x, n_breaks = 6L, labels_week = TRUE)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{group}{A logical indicating if the plot should show groups.  If false
a regrouped plot will be produced.}

\item{stack}{A logical indicating if bars of multiple groups should be
stacked, or displayed side-by-side.}

\item{col_pal}{col_pal The color palette to be used for the groups; defaults
to \code{vibrant} (see \code{?palettes}).}

\item{alpha}{The alpha level for color transparency, with 1 being fully
opaque and 0 fully transparent; defaults to 0.7.}

\item{color}{The color to be used for the borders of the bars; NA for
invisible borders; defaults to \code{NA}.}

\item{xlab}{The label to be used for the x-axis; empty by default.}

\item{ylab}{The label to be used for the y-axis; by default, a label will be
generated automatically according to the time interval used in incidence
computation.}

\item{n_breaks}{n_breaks the ideal number of breaks to be used for the x-axis
labeling}

\item{show_cases}{if \code{TRUE} (default: \code{FALSE}), then each observation will be
colored by a border. The border defaults to a white border unless specified
otherwise. This is normally used outbreaks with a small number of cases.
Note: this can only be used if \code{stack = TRUE}}

\item{border}{If show_cases is TRUE this represents the color used for the
borders of the individual squares plotted (defaults to \code{"white"}).}

\item{na_color}{The colour to plot \code{NA} values in graphs (default: \code{grey}).}

\item{labels_week}{labels_week a logical value indicating whether labels x axis tick
marks are in week format YYYY-Www when plotting weekly incidence; defaults to
TRUE.}

\item{...}{arguments passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}},
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}}, or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, depending
on how the \verb{$date} element is stored in the incidence object.}

\item{fill}{Which variable to colour the fact_plot by.  If NULL (default) no
distinction is made on facet colors.}

\item{legend}{Should a legend accompany the facet_plot}
}
\value{
\itemize{
\item \code{facet_plot()} and \code{plot()} generate a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
\item \code{make_breaks()} a two-element list. The "breaks" element will contain the
evenly-spaced breaks as either dates or numbers and the "labels" element
will contain either a vector of weeks OR a \code{\link[ggplot2:waiver]{ggplot2::waiver()}} object.
\item \code{scale_x_incidence()} a \pkg{ggplot2} "ScaleContinuous" object.
}
}
\description{
incidence2 includes two plotting functions to simplify graph creation.
}
\details{
\itemize{
\item \code{plot} creates a one-pane graph of an incidence object.
\item \code{facet_plot} creates a multi-facet graph of a grouped incidence object.
If the object has no groups it returns the same outout as a call to
\item \code{make_breaks()} calculates breaks from an incidence object that always
align with the bins and start on the first observed incidence.
\item \code{scale_x_incidence()} produces and appropriate \code{ggplot2} scale based on
an incidence object.
}
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist

    inci <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = hospital)

    inci2 <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = c(hospital, gender))

    plot(inci)
    plot(inci, group = FALSE)
    plot(inci, stack = FALSE)
    facet_plot(inci)

    facet_plot(inci2)
    plot(inci2, group = FALSE)
  })
}

}
