% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_labels.R, R/plot.R, R/scale_x_incidence.R
\name{rotate_and_scale}
\alias{rotate_and_scale}
\alias{plot.incidence}
\alias{facet_plot}
\alias{scale_x_incidence}
\title{Rotate and scale incidence plot labels}
\usage{
rotate_and_scale(angle = 0, size = NULL)

\method{plot}{incidence}(
  x,
  fill = NULL,
  stack = TRUE,
  col_pal = vibrant,
  alpha = 0.7,
  color = NA,
  xlab = "",
  ylab = NULL,
  n_breaks = 6,
  show_cases = FALSE,
  border = "white",
  na_color = "grey",
  group_labels = TRUE,
  centre_labels = FALSE,
  legend = c("right", "left", "bottom", "top", "none"),
  ...
)

facet_plot(
  x,
  facets = NULL,
  fill = NULL,
  col_pal = vibrant,
  alpha = 0.7,
  color = NA,
  xlab = "",
  ylab = NULL,
  n_breaks = 3,
  show_cases = FALSE,
  border = "white",
  na_color = "grey",
  group_labels = TRUE,
  centre_labels = FALSE,
  legend = c("bottom", "top", "left", "right", "none"),
  nrow = NULL,
  ...
)

scale_x_incidence(x, n_breaks = 6, group_labels = TRUE, format = NULL, ...)
}
\arguments{
\item{angle}{Angle to rotate x-axis labels (default 90 degrees)}

\item{size}{text size in pts.}

\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{fill}{Which variable to color plots by. If NULL no distinction if made
for plot colors.}

\item{stack}{A logical indicating if bars of multiple groups should be
stacked, or displayed side-by-side. Only used if fill is not NULL.}

\item{col_pal}{col_pal The color palette to be used for the groups; defaults
to \code{vibrant} (see \code{?palettes}).}

\item{alpha}{The alpha level for color transparency, with 1 being fully
opaque and 0 fully transparent; defaults to 0.7.}

\item{color}{The color to be used for the borders of the bars; NA for
invisible borders; defaults to \code{NA}.}

\item{xlab}{The label to be used for the x-axis; empty by default.}

\item{ylab}{The label to be used for the y-axis; by default, a label will be
generated automatically according to the time interval used in incidence
computation.}

\item{n_breaks}{n_breaks the ideal number of breaks to be used for the x-axis
labeling}

\item{show_cases}{if \code{TRUE} (default: \code{FALSE}), then each observation will be
colored by a border. The border defaults to a white border unless specified
otherwise. This is normally used outbreaks with a small number of cases.
Note: this can only be used if \code{stack = TRUE}}

\item{border}{If show_cases is TRUE this represents the color used for the
borders of the individual squares plotted (defaults to \code{"white"}).}

\item{na_color}{The colour to plot \code{NA} values in graphs (default: \code{grey}).}

\item{group_labels}{group_labels a logical value indicating whether labels x axis tick
marks are in week format YYYY-Www when plotting weekly incidence; defaults to
TRUE.}

\item{centre_labels}{Sgould labels on the axis be centred on the bars. This
only applies to intervals that produce unambiguous labels (i.e \verb{1 day},
\verb{1 month}, \verb{1 quarter} or \verb{1 year}).  Defaults to \code{FALSE}.}

\item{legend}{Position of legend in plot.}

\item{...}{arguments passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}},
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_datetime()}}, or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}},
depending on how the \verb{$date} element is stored in the incidence object.}

\item{facets}{Which variable to facet plots by.  If NULL will use all
group_labels of the incidence object.}

\item{nrow}{Number of rows.}

\item{format}{Character string of desired format.  See \code{?strptime}.}
}
\value{
\itemize{
\item \code{facet_plot()} and \code{plot()} generate a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
}
\description{
incidence2 includes two plotting functions to simplify graph creation.
}
\details{
\itemize{
\item \code{plot} creates a one-pane graph of an incidence object.
\item \code{facet_plot} creates a multi-facet graph of a grouped incidence object.
If the object has no groups it returns the same outout as a call to
}
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist

    inci <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = hospital)

    inci2 <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = c(hospital, gender))

    plot(inci)
    plot(inci, fill = hospital)
    plot(inci, fill = hospital, stack = FALSE)

    facet_plot(inci)
    facet_plot(inci2)
    facet_plot(inci2, facets = gender)
    facet_plot(inci2, facets = hospital, fill = gender)
  })
}

}
