% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.incidence2}
\alias{plot.incidence2}
\alias{facet_plot}
\alias{facet_plot.incidence2}
\title{Plotting functions}
\usage{
\method{plot}{incidence2}(
  x,
  fill = NULL,
  stack = TRUE,
  title = NULL,
  col_pal = vibrant,
  alpha = 0.7,
  color = NA,
  xlab = "",
  ylab = NULL,
  n_breaks = 5,
  show_cases = FALSE,
  border = "white",
  na_color = "grey",
  legend = c("right", "left", "bottom", "top", "none"),
  angle = 0,
  size = NULL,
  ...
)

facet_plot(x, ...)

\method{facet_plot}{incidence2}(
  x,
  facets = NULL,
  stack = TRUE,
  fill = NULL,
  title = NULL,
  col_pal = vibrant,
  alpha = 0.7,
  color = NA,
  xlab = "",
  ylab = NULL,
  n_breaks = 3,
  show_cases = FALSE,
  border = "white",
  na_color = "grey",
  legend = c("bottom", "top", "left", "right", "none"),
  angle = 0,
  size = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{fill}{Which variable to color plots by. If NULL no distinction if made
for plot colors.}

\item{stack}{A logical indicating if bars of multiple groups should be
stacked, or displayed side-by-side. Only used if fill is not NULL.}

\item{title}{Optional title for the graph.}

\item{col_pal}{col_pal The color palette to be used for the groups; defaults
to \code{vibrant} (see \code{?palettes}).}

\item{alpha}{The alpha level for color transparency, with 1 being fully
opaque and 0 fully transparent; defaults to 0.7.}

\item{color}{The color to be used for the borders of the bars; NA for
invisible borders; defaults to \code{NA}.}

\item{xlab}{The label to be used for the x-axis; empty by default.}

\item{ylab}{The label to be used for the y-axis; by default, a label will be
generated automatically according to the time interval used in incidence
computation.}

\item{n_breaks}{n_breaks the ideal number of breaks to be used for the x-axis
labeling}

\item{show_cases}{if \code{TRUE} (default: \code{FALSE}), then each observation will be
colored by a border. The border defaults to a white border unless specified
otherwise. This is normally used outbreaks with a small number of cases.
Note: this can only be used if \code{stack = TRUE}}

\item{border}{If show_cases is TRUE this represents the color used for the
borders of the individual squares plotted (defaults to \code{"white"}).}

\item{na_color}{The colour to plot \code{NA} values in graphs (default: \code{grey}).}

\item{legend}{Position of legend in plot.}

\item{angle}{Rotation angle for text.}

\item{...}{other arguments to pass to \code{\link[=scale_x_incidence]{scale_x_incidence()}}.}

\item{facets}{Which variable to facet plots by.  If NULL will use all
group_labels of the incidence object.}

\item{nrow}{Number of rows.}
}
\value{
\itemize{
\item \code{facet_plot()} and \code{plot()} generate a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
}
\description{
incidence2 includes two plotting functions to simplify graph creation.
}
\details{
\itemize{
\item \code{plot} creates a one-pane graph of an incidence object.
\item \code{facet_plot} creates a multi-facet graph of a grouped incidence object.
If the object has no groups it returns the same output as a call to
\code{\link[=plot]{plot()}}.
\item If the \code{\link[=incidence]{incidence()}} object has a rolling average column then that
average will be overlaid on top.
}
}
\examples{
if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist

    inci <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = hospital)

    inci2 <- incidence(dat,
                      date_index = date_of_onset,
                      interval = 7,
                      groups = c(hospital, gender))

    plot(inci)
    plot(inci, fill = hospital)
    plot(inci, fill = hospital, stack = FALSE)

    facet_plot(inci)
    facet_plot(inci2)
    facet_plot(inci2, facets = gender)
    facet_plot(inci2, facets = hospital, fill = gender)
  })
}
}
