% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool}
\alias{pool}
\title{Pool 'incidence' across groups}
\usage{
pool(x, groups = NULL)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{groups}{The groups to sum over.  If \code{NULL} (default) then the function
ignores all groups.}
}
\description{
This function pools incidence across specified groups of an \code{incidence}
object. The resulting \code{\link[=incidence]{incidence()}} object will contains counts summed over
the groups present in the input.
}
\seealso{
The \code{\link[=incidence]{incidence()}} function to generate the 'incidence'
objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
