% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R, R/dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.table}
\alias{as_tibble}
\alias{mutate}
\alias{nest}
\alias{select}
\alias{summarise}
\alias{unpack}
\alias{unnest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyr}{\code{\link[tidyr]{nest}}, \code{\link[tidyr]{unnest}}, \code{\link[tidyr:pack]{unpack}}}
}}

