% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_interval}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grates}{\code{\link[grates:grate_accessors]{get_interval}}}
}}

