% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_average.R
\name{rolling_average}
\alias{rolling_average}
\title{Add a rolling average}
\usage{
rolling_average(x, before = 2)
}
\arguments{
\item{x}{An \code{\link[=incidence]{incidence()}} object.}

\item{before}{how many prior dates to group with.}
}
\value{
A \code{\link[=incidence]{incidence()}} object with an additional \code{rolling_average} column.
}
\description{
Add a rolling average
}
\note{
If groups are present the average will be calculated across each
grouping, therefore care is required when plotting.
}
\examples{
data(ebola_sim_clean, package = "outbreaks")
dat <- ebola_sim_clean$linelist

inci <- incidence(dat,
                  date_index = date_of_onset,
                  interval = "week",
                  last_date = "2014-10-05",
                  groups = gender)

inci \%>\%
  regroup() \%>\%
  rolling_average(before = 2) \%>\%
  plot(color = "white")

inci \%>\%
  rolling_average(before = 2) \%>\%
  facet_plot(color = "white")

}
