% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.incidence}
\alias{summary.incidence}
\title{Summary of an incidence object}
\usage{
\method{summary}{incidence}(x, ...)
}
\arguments{
\item{x}{An 'incidence' object.}

\item{...}{Not used.}
}
\value{
x (invisibly).
}
\description{
Summary of an incidence object
}
